move more functionality into server/
