+++ grunt-helpers.js
@@ -11,7 +11,7 @@ var _ = require("underscore")
   , prompt = require('prompt')
   , exec = require('child_process').exec
   , spawn = require('win-spawn')
-  , parser = require('./lib/parser')
+  , parser = require('./lib/server/parser.js')
   , namp = require('namp')
   , parseXmlString = require('xml2js').parseString
   , appiumVer = require('./package.json').version
