+++ lib/android.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var errors = require('./errors')
+var errors = require('./server/errors.js')
   , adb = require('../android/adb')
   , _ = require('underscore')
   , logger = require('./server/logger.js').get('appium')
