+++ lib/appium.js
@@ -1,7 +1,7 @@
 // Appium webserver controller methods
 // https://github.com/hugs/appium/blob/master/appium/appium.py
 "use strict";
-var routing = require('./routing')
+var routing = require('./server/routing.js')
   , loggerjs = require('./server/logger.js')
   , logger = loggerjs.get('appium')
   , setLogFile = loggerjs.setLogFile
