+++ lib/bin.js
@@ -5,7 +5,7 @@ var net = require('net')
   , repl = require('repl')
   , colors = require('colors')
   , appium  = require('../server')
-  , parser = require('./parser');
+  , parser = require('./server/parser.js');
 
 var startRepl = function() {
   var help = function() {
