+++ lib/device.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var errors = require('./errors')
+var errors = require('./server/errors.js')
   , request = require('request')
   , _ = require('underscore')
   , exec = require('child_process').exec
