+++ lib/firefoxos.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var errors = require('./errors')
+var errors = require('./server/errors.js')
   , _ = require('underscore')
   , logger = require('./server/logger.js').get('appium')
   , net = require('net')
