+++ lib/ios.js
@@ -18,7 +18,7 @@ var path = require('path')
   , rotateImage = helpers.rotateImage
   , rd = require('./hybrid/ios/remote-debugger')
   , wkrd = require('./hybrid/ios/webkit-remote-debugger')
-  , errors = require('./errors')
+  , errors = require('./server/errors.js')
   , deviceCommon = require('./device')
   , status = require("./uiauto/lib/status")
   , IDevice = require('node-idevice')
