+++ lib/server/controller.js
@@ -1,14 +1,14 @@
 // Appium webserver controller methods
 // https://github.com/hugs/appium/blob/master/appium/server.py
 "use strict";
-var status = require('./uiauto/lib/status')
-  , logger = require('./server/logger.js').get('appium')
+var status = require('../uiauto/lib/status.js')
+  , logger = require('./logger.js').get('appium')
   , _s = require("underscore.string")
   , swig = require('swig')
   , path = require('path')
-  , version = require('../package.json').version
-  , proxy = require('./proxy')
-  , responses = require('./responses')
+  , version = require('../../package.json').version
+  , proxy = require('./proxy.js')
+  , responses = require('./responses.js')
   , getResponseHandler = responses.getResponseHandler
   , respondError = responses.respondError
   , respondSuccess = responses.respondSuccess
@@ -983,6 +983,6 @@ exports.guineaPig = function(req, res) {
 };
 
 exports.getTemplate = function(templateName) {
-  return swig.compileFile(path.resolve(__dirname, "server", "templates",
+  return swig.compileFile(path.resolve(__dirname, "templates",
         templateName + ".html"));
 };
