+++ lib/server/errors.js
