+++ lib/server/main.js
@@ -11,7 +11,7 @@ var http = require('http')
   , appiumRev = null
   , async = require('async')
   , _ = require("underscore")
-  , parser = require('../parser')
+  , parser = require('./parser.js')
   , io = require('socket.io')
   , gridRegister = require('../../grid_register');
