+++ lib/server/parser.js
@@ -1,6 +1,6 @@
 "use strict";
 var ap = require('argparse').ArgumentParser
-  , pkgObj = require('../package')
+  , pkgObj = require('../../package.json')
   , path = require('path')
   , _ = require('underscore');
