+++ lib/server/proxy.js
@@ -1,9 +1,9 @@
 "use strict";
 
 var _s = require('underscore.string')
-  , logger = require('./server/logger.js').get('appium')
-  , doRequest = require('./device').doRequest
-  , respondError = require('./responses').respondError
+  , logger = require('./logger.js').get('appium')
+  , doRequest = require('../device.js').doRequest
+  , respondError = require('./responses.js').respondError
   , _ = require('underscore');
