+++ lib/server/responses.js
@@ -1,7 +1,7 @@
 "use strict";
 
-var logger = require('./server/logger.js').get('appium')
-  , status = require('./uiauto/lib/status')
+var logger = require('./logger.js').get('appium')
+  , status = require('../uiauto/lib/status.js')
   , _ = require('underscore');
 
 var getSessionId = function(req, response) {
