+++ lib/server/routing.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var controller = require('./controller');
+var controller = require('./controller.js');
 
 module.exports = function(appium) {
   var rest = appium.rest;
